+++ ChangeLog.md
@@ -1,6 +1,12 @@
 URI handling change log
 =======================
 
+## ?.?.? / ????-??-??
+
+* Implemented read access to parameters via `params()` and `param()`
+  methods on the `URI` class
+  (@thekid)
+
 ## 0.1.0 / 2017-03-24
 
 * Hello World! First release - @thekid